Broke out Bullpen and SOD endpoints into separate controllers
